Bootstore: Start implementing coordinator
This commit introduces the coordinator for the bootstore 2PC protocol.
Currently only the rack initialization part of the coordinator is implemented.
Reconfiguration is coming later.

A property based test was written that shows a single rack initialization that works perfectly.
This is good for demonstration purposes, but a more advanced stateful test will be written in the future
to add failures / dropped messages when reconfiguration is implemented.

A smattering of other changes are here as well.

* Sqlite connections now live inside a RefCell to allow logically immutable
methods.
* Shares are now zeroized on drop
* Trust quorum membership is now a BTreeSet of certs rather than a Vec