Set instance state to failed on bad instance_put
If a call to sled-agent's `instance_put` fails with anything other than
a 400 (bad request), set instance's state to failed - we cannot know
what state the instance is in.

This exposed another two problems:

- Nexus would only delete network interfaces off instances that were stopped
- Nexus would only detach disks off instances that were creating or stopped.

This commit changes the relevant network interface queries to allow
deletion of network interfaces for instances that are either stopped or
failed, and adds Failed to the list of ok_to_detach_instance_states for
disks. Note that network interface insertion still requires an instance
to be stopped, not failed.

Closes https://github.com/oxidecomputer/omicron/issues/1713