Cleanup Crucible resources with volume delete saga (#1706)
Add a volume delete saga, and call that whenever a volume needs to be
deleted. Crucible resources now have accounting for their use, and when
the number of references reaches zero then cleanup occurs.

Adds a table to store created region snapshots as part of this
accounting, and make a record of when a Crucible snapshot is taken of a
region. Region snapshots are uniquely identified by (dataset_id,
region_id, snapshot_id), and additionally contain a snapshot address for
identification as part of a volume construction request.

All DELETE calls to Crucible Agent(s) are now only performed by the
volume delete saga.

Add tests for cases where if Nexus is properly accounting for snapshots.
It should not delete the disk's corresponding region when the disk is
deleted because a snapshot exists, or if multiple disks reference a
running snapshot.

A small change to the simulated Crucible agent is also included in this
commit - logic was moved out of the HTTP endpoint function into the
Crucible struct. Logic doesn't belong in HTTP endpoint functions :)

Closes https://github.com/oxidecomputer/omicron/issues/1632