Improve `Screen` trait
Draw methods are no longer allowed to mutate self. This was achieved
by adding a provided `resize` method that performs any necessary
state mutations for drawing, and which is called independently of
`draw` as necessary.

When we create a `ScreenButtonState`, we don't actually know
where it's going to be placed because the constructor doesn't have
access to the frame size. With the aformentioned changes we can now
simplify the constructor and place it when `resize` is called.