Make sure # of sectors of generated cloud-init configdrive is multiple of sectors-per-track.
The mtools suite that cloudbase-init uses is more of a stickler and
requires that the number of sectors of the FAT fs is a multiple of its
sector-per-track value. The in-memory configdrive we were creating fell
afoul of this check. Given we're creating a small disk we can make sure
this holds trivially.