Pull in updated fats-rs crate
The volume label for a FAT filesystem may be stored in 2 different
locations (the boot sector and a special entry in the root dir).
Older versions of the fatfs crate were only setting the label in the
former but cloudbase-init (via mlabel from mtools) was looking for the
label in the latter spot. Not finding it there it would fail to identify
the drive as a valid configdrive for the cloud-init metadata.