Add Windows Server 2022 image.
This was built using the Windows Server 2022 Evaluation ISO. It has the
VirtIO block and network drivers pre-installed. It also includes
cloudbase-init and is preconfigured to work with the metadata we expose
via the in-memory FAT drive. That allows it configure the "oxide" user
to accept the SSH passed in from the control plane. In addition to SSH,
RDP is also enabled for the image (though you will need to change the
password from the randomly generated one via SSH `net user oxide *`).