Wicket: Add update support code required for MUP
For MUP, users will write a TOML file similar to
`example-rack-update-spec.toml` that maps to a `RackUpdateSpec`. Users
will then create a release update via running `mupdate-pkg`. This package
will be sent via ssh to wicket which will unpack it via `Manifest::load`
and display the contents to the user. Wicket will use uploaded update packages
to update the rack with the help of MGS and other downstream services.

Note that the format of an update manifest and package is homeade here.
We may possibly re-use it with a TUF wrapper in the future, or discard it
altogether post MVP.