Teach Nexus about the Pantry
Boot the new Pantry service zone, and use that to snapshot disks that
are not attached to a currently running propolis. This functionality has
to be driven in Nexus as part of a Saga because using the Pantry
requires multiple steps (each one of which can fail). If this was simply
a matter of proxying a call, the sled agent would still be used.

A simulated Pantry was implemented for the unit tests, and the logic for
simulating a snapshot of an unattached disk moved there. The server code
and Pantry interface were wholesale copied from the Crucible repo - it'd
be nice to have a way to use the same interface(s) here and in Crucible
(for the simulated agent too!) but I don't know a good way to share
these cross-repo.

When the sled-agent brings up service zones, it inserts corresponding
DNS records into the AZ's internal DNS server, and creating a Pantry
client (or any service zone client) will need to query this internal DNS
server for those records. This means that the simulated sled agent needs
to also start an internal DNS server and insert records into it, which
this commit implements. There's some minor refactoring in the
internal-dns crate to create a `start` method for this purpose.

This means that Nexus needs to store the Resolver, and needs a way to
update the Resoler that is used during integration tests. Importantly,
services launched as part of integration tests will not use a standard
known port (otherwise there would be many collisions), and a way to
resolve both SRV (to get the port) and AAAA (to get the IP) records is
required.

This commit also bumps the crucible and propolis git revs in the
respective Cargo.toml files to the latest.