Add Windows image (#1850)
This adds a new image to the populate-images script based on `Windows Server 2022 Standard Evaluation (Desktop Experience)` ([scripts used](https://github.com/oxidecomputer/windows-image-builder)).

It has the VirtIO network driver installed so guest networking works (both inbound & outbound). While we use NVMe for storage, it also has the VirtIO block driver installed which is needed for the cloud-init configdrive we expose to guests. Similarly, it also comes with [cloudbase-init](https://cloudbase-init.readthedocs.io/en/latest/intro.html) installed and configured to work with the cloud-init metadata we provide. (There were a few small tweaks made to the in-memory FAT drive we generate for the metadata to get it working.)

First boot can take a few minutes as it begins up in a state as if you had just left the initial Windows Setup. It will skip past the Out of Box Experience (OOBE). After a reboot or two it'll be ready to SSH into. (Subsequent boots will be faster.)

By default it configures an admin user "oxide" with a random password, copies over the SSH key provided by the control plane. It will also update the hostname to match the control plane value. It will also automatically extend the OS partition size (which is ~14G in the base image) to match the configured disk size. RDP is also enabled but you must configure a password first which can be done over SSH by running `net user oxide *`.

This also pulls in a rev of propolis (& crucible) with the updated bootrom.