Introduce wicketd
`wicketd` is the backend service that wicket will interact with to
provide technician port based operator functionality to the rack. This
is the start of an implementation and there is much more to do.
Currently, `wicketd` is essentially a caching MGS proxy for the `/sp`
endpoint of MGS. The MGS endpoint is paginated, but for wicket, we
don't really mind waiting as we plan to poll for changes periodically.
To ensure we get all information, we just attempt to retrieve the max
amount of components (36) within a given timeout. This limits us to one
page that wicket then stores in memory and serves as `/inventory`.

There is an open TODO to fetch the SP inventory for each SP retrieved
from `/sp` via the MGS `/sp/{type}/{slot}/component` endpoint. Once
this is implemented, wicket will be modified to talk to wicketd
and serve this inventory instead of the fake inventory currently there.

Since we want to allow proxying information from MGS with a minimum
of hassle, `gateway-client` was changed to derive `schemars::JsonSchema`
for generated types so those could be served from `wicketd`.

Since this is a caching proxy for inventory, when wicket users request
inventory a request gets forwarded over a channel to the `MgsManager`
which returns an `Arc<RackV1Inventory>`. This prevents the need to copy
this data. When the inventory gets updated a new `Arc<RackV1Inventory>`
is created and will be served over the same channel. This channel
based mechanism works nicely, but it's worth noting the alternative of
using an `Arc<Mutex<` or `<Arc<RwLock`, or even a
`tokio::sync::watch::channel`. There may be performance tradeoffs with
them, but the behavior is generally the same, and I find the message
passing flow easier to follow in most cases. I'm open to changing this
down the road.

There is also an intention to support ETAGs for the inventory, which
would negate the need to bother with an `Arc` when sending the message
over a channel, as copying would be so rare. This is left for a follow
up commit, although there is some embedded support.

It should also be noted that polling the SPs every 10 seconds and
retrieving inventory is really inefficient. Instead MGS should maintain
a generation count of when inventory changes and provide information
about what changed. Alternatively, MGS could also use ETAGs and just
change the ETAG for `/sp` when anything changes. This would have more
overhead, but *may* be simpler to implement. In order to implement
either of these optimizations however, there would need to be some SP
support for keeping track of changes. It's worth trying to determine
when inventory can even change. Is it during upgrades or insertion of
new hardware only?

This code was tested manually by running the following in separate
terminals:
 * SP simulators (2 Sleds, 1 Switch)
 * MGS
 * wicketd
 * curl

```
➜  omicron git:(dfc01324) ✗ cargo run --bin sp-sim -- sp-sim/examples/config.toml
```

```
➜  omicron git:(dfc01324) ✗ cargo run --bin mgs -- run --id 14c8ad58-c55a-430c-9db5-5db2bfc3780e --address '[::1]:9999' gateway/examples/config.toml`
```

```
➜  omicron git:(dfc01324) ✗ cargo run -p wicketd -- run --address "[::1]:8000" wicketd/examples/config.toml
```

```
➜  ~ curl --silent "[::1]:8000/inventory" | jq
{
  "sps": [
    [
      {
        "details": {
          "state": "enabled",
          "serial_number": "00000000000000000000000000000002"
        },
        "info": {
          "details": {
            "present": "yes",
            "ctrl_detect_0": true,
            "ctrl_detect_1": false,
            "flt_a2": false,
            "flt_a3": false,
            "flt_rot": false,
            "flt_sp": false,
            "id": 17,
            "power": true
          },
          "id": {
            "slot": 0,
            "type": "sled"
          }
        }
      },
      []
    ],
    [
      {
        "details": {
          "state": "enabled",
          "serial_number": "00000000000000000000000000000001"
        },
        "info": {
          "details": {
            "present": "yes",
            "ctrl_detect_0": true,
            "ctrl_detect_1": false,
            "flt_a2": false,
            "flt_a3": false,
            "flt_rot": false,
            "flt_sp": false,
            "id": 18,
            "power": true
          },
          "id": {
            "slot": 0,
            "type": "switch"
          }
        }
      },
      []
    ],
    [
      {
        "details": {
          "state": "enabled",
          "serial_number": "00000000000000000000000000000003"
        },
        "info": {
          "details": {
            "present": "yes",
            "ctrl_detect_0": true,
            "ctrl_detect_1": false,
            "flt_a2": false,
            "flt_a3": false,
            "flt_rot": false,
            "flt_sp": false,
            "id": 17,
            "power": true
          },
          "id": {
            "slot": 1,
            "type": "sled"
          }
        }
      },
      []
    ]
  ]
}
```