Wait for instance_ensure too
Booting instances with a lot of ram takes longer than the default 15s
timeout that progenitor uses. Create a custom client (just like #1888
does) that waits indefinitely.

Also, print the result of the instance_ensure call, which was essential
during debugging.

Before this commit, instances with 32 GB of ram would time out and cause
the propolis zone to be torn down, even though propolis was reporting
success. With this commit, instances with as large as 64 GB of ram were
tested and booted ok.