[gateway]: Remove pagination for '/sp' route (#1921)
This commit removes pagination for /sp as detailed in #1913. The only thing done differently is that I kept the timeout. In tests that seems to cause 5 UDP retries according to the logs. We can reconsider a change here, but maybe in a follow up.

This also fixes the gateway tests to utilize the changed API and uses it in wicketd.