[wicket] Start to pull real inventory from wicketd
This commit hooks wicket up to wicketd via wicketd-client. Wicket no
longer publishes fake inventory, and no longer plans to interact with
MGS. Instead it only talks to wicketd which acts as a caching proxy for
MGS inventory.

The data retrieved from wicketd is converted into types already used
by the display screen infrastructure in wicket. This is what's happening
in `inventory.rs`. When the inventory is updated the screen is redrawn.

This is just a start, and quite a few things are missing:
 * Real power state - this is currently faked to A2 for everything
 * ComponentInfo (actual SP inventory)
 * Version info - This still needs to be implemented on the SPs and
   plumbed all the way up
 * A display for the last time wicketd was talked to and if MGS is
   actually alive. We want to show how stale our data is. Right now
   if wicketd goes down we'll log it but not notice in the UI. If MGS
   goes down we don't have any idea.
 * ETAG support and diff support so we don't redraw the same inventory
   over and over.

This was tested via running MGS, wicketd and sp simulators on my local
helios box. If you run wicket standalone it still works fine, but the
component screen won't show any inventory.