[wicket/wicketd] Retrieve SpComponentList (#1953)
Add SpComponentInfo to RackV1Inventory by pulling it from MGS as
necessary. We only pull from MGS when the state of the SP has changed
or we do not yet have the component info. This is because the info is
static except for when the SP is upgraded.

We then expose this information via the wicketd-client with the
appropriately derived traits and plumb it up through wicket.