move xde onu to stlouis onu
We are trying to migrate workflows from the xde and integration branches
to the stlouis branch. This commit updates the opte install script to
grab the stlouis onu/p5p build from the os-build repo instead of the xde
branch build.

I've tested this by setting up a RAP environment and the OPTE/XDE
plumbing comes up successfully.