Update lookup fns introduced in the V1 API to return `lookup::*` over `authz::*` (#2042)
This PR is mostly in response to @davepacheco's feedback on #1957.
Special thanks to @smklein for helping work through some rust lifetime
issues here.

Previously the `instance_lookup` and `project_lookup` functions returned
a `authz::Instance` or `authz::Project` respectively that was authorized
with read permissions. The challenge here is that this meant that in
several places we required duplicate lookups to be able to resolve the
instance with the right permissions. To avoid that we're passing around
the lookup step itself instead.