MGS: Import `Communicator` and supporting code from mgs repo (#2053)
There are no meaningful changes in this PR, other than touching up
import paths and type names from the move. This is just importing code
that was moved out to the MGS repo in #1809 (which moved the
`gateway-messages` and `gateway-sp-comms` crates in their entirety). The
bits being moved back here are directly relevant to MGS proper (e.g.,
RFD 250-style discovery over all the VLAN interfaces), and are not
directly relevant to the communication protocol between MGS and one SP.