MGS: Don't use ignition results to decide whether to contact an SP
This PR changes both the "get state of an SP" and the "get state of all
SPs" to, primarily, no longer first ask the ignition controller whether
the SP on skip talking to the SP if ignition indicates it is not.
Instead, this changes `SpInfo` such that it can now indicate ignition
status/failure and SP status/failure independently.

Additionally for the "get state of all SPs" endpoint, because we no
longer use ignition as a filter, we no longer need to take a timeout;
instead, we send requests to all SPs concurrently and wait for them all
to respond with our normal rules (i.e., our configured number of
attempts / timeout per attempt for individual requests).