Include extent information in Crucible VolumeConstructionRequests (#2082)
Pick up the latest definition of
`crucible_client_types::VolumeConstructionRequest` and the corresponding
Propolis bits. The new definition requires a
`VolumeConstructionRequest::Region` to specify the number of extents in
the region and the number of blocks in each extent. The disk creation
saga already has all this data (it needs it to ask Crucible to create
the regions in the first place), so this is just a matter of plumbing
that data into the serialized VCR that Nexus saves in its database.