Add `identity()` method for `AssetIdentityMetadata`, move def to another file (#2102)
I noticed two things while working on #2100.

1. `AssetIdentityMetadata` is defined in `views.rs`, but it is used by
both DB models and views. To me that makes it not primarily a view, so
it should go in a different file. I moved it to
`nexus/types/src/identity.rs`, where `Asset` is defined.
2. Instead of the `identity()` method for copying around the identity
metadata (which `Resource` has), there is a `impl From<T: Asset> for
AssetIdentityMetadata`, which is perfectly fine, but I think it's nicer
to have them be the same.

As far as I remember, these are accidents of history and not for any
particular reason.