[installinator] initial version (#2027)
Implement the initial version of the "installinator" described in [RFD
345](https://rfd.shared.oxide.computer/rfd/0345#_host_phase_2_install).

Per the RFD, the installinator is "[t]he key process in the Phase-2
Trampoline Image that fetches data over the bootstrap network to install
on an M-2 device."

> 2.6 Host Phase-2 Install
>
> At the start of this step, the host has booted, it’s running user-land
processes, the bootstrap network (and not the underlay network) is
available (meaning that T6 and Maghemite are both up). At the conclusion
of this step, the Phase-2 Host Image will be present on an M.2 device
(of the two).
>
> The Trampoline image contains a single-purpose program called
Installinator that downloads the Phase-2 Host Image and places it on an
M.2. Installinator discovers peers on the bootstrap network which
includes sled-agent and wicketd instances; it connects to each and
requests the desired Phase-2 Host Image by ID and/or checksum.
>
> Every sled-agent serves data from its own, current M.2; wicketd serves
additional images that have been uploaded to it.

This isn't complete yet; the TODOs are tracked in
https://github.com/oxidecomputer/omicron/issues/2061.

Questions:

- [x] Is it OK to buffer the downloaded artifact to memory?
- Answer is *yes* for V1. ~~In the future, we might want to write
directly out to disk as we verify it (assuming we do streaming
verification).~~ (This is not an issue for a long time because we have
orders of magnitude more RAM than the size of the host image).
- [x] Currently we attempt fetches from all peers and take the first one
that can serve us the entire artifact. Do we need to first check for the
presence of the artifact then attempt fetches one-by-one? (This is going
to depend on how often the installinator runs and how many peers are
expected to have the artifact, I guess).
  - Resolved in retry/fetch loop.