MGS: Remove `Communicator` (#2112)
This was a helper type that used to do more, but has evolved into little
more than a glorified wrapper around `ManagementSwitch`. This commit
removes `Communicator` entirely in favor of using `ManagementSwitch`
directly. It has a handful of related cleanup changes: renaming some
methods for clarity, making some methods private, and a minor amount of
error handling cleanup.

Builds on #2103 and should be merged after it.