Add a Nexus internal API endpoint for disk remove read only parent
Created a new internal API endpoint that allows a caller to remove
a read only parent from a volume if they only know the UUID of the
disk associated with that volume.  This is basically just a wrapper
that takes a disk UUID, extracts the volume_id from that disk, then
calls the volume saga to remove the read only parent.