Add a Nexus internal API endpoint for disk remove read only parent (#2157)
When we create a disk, the saga generates both a disk_id and a
volume_id We use both ids when we create the disk database record
However, the `volume_id` here is only used as the key to store the
volume data in the database. That volume_id is never exposed to
anyone outside Nexus.

For the actual volume data structure and later the 
VolumeConstructionRequest, we use the value in the disk_id, but 
call it volume_id.

However, Propolis (who needs to call Nexus and tell it to remove a read
only parent only has the top level disk UUID. The internal volume_id is not
exposed.

This adds a new internal API endpoint that will take the disk UUID,
then lookup the volume_id and then call the volume layer remove 
read only parent saga. Adding a layer like this also allows tests that
operate only on a volume level to not require any changes.

Co-authored-by: Alan Hanson <alan@oxide.computer>