Use strings for update artifact versions (#2182)
This is part 1 for getting TUF repo generation done; this changes a
common data structure for updates so I can do some other data type
shifting on top of it.

I had initially used integer versions for update artifacts as a bit of a
strawman. Over the course of a few meetings on updates that culminated
in [RFD 334](https://rfd.shared.oxide.computer/rfd/0334), still in
prediscussion, I've been convinced that we need something a bit more
nuanced than a monotonically-increasing integer. While RFD 334 still
shows a determination that a monotonically-increasing integer is okay,
the inability to describe an "in-between" version (for the
certainly-likely future case of producing a bug fix on an older,
still-supported branch while customers move to the latest branch) is
unacceptable.

Semver is the likely version format, but as far as Nexus and Sled Agent
talking to each other is concerned, the version could be any string we
like as long as it's unique for that name and artifact kind. (Strings
also keep the use of `#[derive(JsonSchema)]` simple.)