Update to latest MGS with a single shared UDP socket (#2191)
Most of the changes in this PR are miscellaneous fallout from minor type
/ API changes made in MGS over the past few weeks. The most substantial
changes are in `ManagementSwitch::new()`, which now either creates a
`SharedSocket` (for real deployments) or a UDP socket per SP (for sim
deployments and tests). This is a little messy but seemed like the most
straightforward way to continue to support running MGS proper without
being in a real switch zone.