[wicket] add a state to indicate inventory is unavailable (#2194)
Currently, the inventory being unavailable is indicated by returning
an empty list. While this works for now; in the future, we'd like to be
able to add heartbeat functionality. This is so that wicketd can
return the last time it talked to MGS. With this design, the duration
would naturally fall under the new `GetInventoryResponse::Inventory`
variant.