[MGS] sled-agent launches and reconfigures MGS in the switch zone (#2212)
There are three related changes in this PR:

1. MGS can now start multiple dropshot servers (that all share the same
server context), allowing it to listen on multiple addresses. In
particular for us, this includes localhost (for talking to wicketd
before the rack is setup) and the underlay network (once it exists, for
talking to nexus). The addresses on which MGS is listening can be
changed while MGS is running.
2. MGS can now read its own SMF properties (there are only two: the UUID
identifying our MGS instance, and the list of addresses on which to
listen). If sent `SIGUSR1`, MGS will reread its SMF properties and
change its listening addresses if they've changed. (Changes to the ID
are not allowed.) MGS's SMF manifest now includes `:kill -USR1` as its
refresh method, allowing `smf refresh mgs` to work.
3. `sled-agent` sets the relevant properties for MGS to listen on
localhost (always) and the underlay network (when it exists).

In combination with other recent changes to MGS (particularly #2207),
this PR should allow sled-agent to successfully launch and configure MGS
within a real switch zone.

Co-authored-by: Sean Klein <sean@oxide.computer>