Add support for new tofino device location (#2179)
The tofino device will be moving from /dev/tofino to /dev/tofino/[0-9]+,
to support a world in which we may have multiple tofino devices attached
to a single scrimlet. For the foreseeable future we will continue to
have only a single device, so this change will cause sled-agent to look
for the first device in that namespace.