[wicket] add a status bar with liveness information for wicketd and MGS (#2195)
Within wicket, it's important to know if we're seeing outdated
information. Add a status bar widget, and use it to display liveness
information for both wicketd and MGS.

In the future, we can use the status bar to show information about
updates and other information.

Here's what it looks like when it's live:


![image](https://user-images.githubusercontent.com/180618/213831523-58e1bcb2-b1a5-4474-b9c6-c3175073e79b.png)

And here's when it hasn't received information from wicketd for a little
while:


![image](https://user-images.githubusercontent.com/180618/213831540-95a2180a-55c3-4d6e-a812-62fe30ac88ea.png)

Depends on #2193 and #2194.