[MGS] Bring gateway-cli closer to parity with faux-mgs (#2233)
This PR reworks `gateway-cli` (which had languished quite a bit) to be
closer to in parity with `faux-mgs`, with a few notable exceptions:

1. The SP the user wants to talk to is specified in terms of a logical /
physical identifier (e.g., `sled/7` is "the gimlet in slot 7") instead
of an IP address or interface. The only network connection made by
gateway-cli itself is to MGS.
2. There is no `discover` subcommand. MGS uses discovery packets to find
and map out online SPs, but doesn't expose those to its clients.
3. The `IgnitionLinkEvents` and `ClearIgnitionLinkEvents` commands are
omitted. My current understanding is that these are for manual ignition
debugging only and not events we want the control plane to care about in
general. If that changes, they're in the MGS protocol already and would
need endpoints added.
4. A few commands have no MGS endpoint; these are left as `todo!()`s
that I will address in followup PRs that touch both MGS proper and
`gateway-cli`.