[wicketd] fix delayed status of updates from MGS (#2241)
Previously, we were only sending notifications over this channel if the
inventory had changed. This meant that we weren't getting newer instants
if the inventory was unchanged, causing wicket to say "delayed" rather
than "live".

Tested by setting up MGS and sp-sim, then ensuring that the status was
always "live".