Create separate underlay and bootstrap etherstubs (#2309)
VNICs used across zones shared a single etherstub. This commit changes
it so that the bootstrap and underlay networks no longer share an
etherstub, and therefore are no longer on the same L2 network.

All service zones have VNICs on the underlay network. A new VNIC for the
bootstrap network will be added to the switch network in a follow up.

Fixes https://github.com/oxidecomputer/omicron/issues/2297