[MGS] Use dropshot's web socket support for serial console endpoint (#2316)
The primary driver for this PR was removing the `TODO` in gateway-cli
where the progenitor MGS client didn't know that the serial console
endpoint was a websocket, so we had to manually connect via
`tokio_tungstenite`. As a very nice side effect, we get to drop all the
manual websocket upgrade / HTTP header stuff we had previously.