Add Bootstrap vnic and address to switch zone (#2320)
Add a bootstrap vnic (over the bootstrap etherstub in the global zone)
to the switch zone, along with a bootstrap IP address that can be
listened on by the wicketd artifact server. The wicketd artifact server
is used for serving OS and host images, and a follow up change will be
made for it to listen on the bootstrap server.

Currently we don't expect any other zones besides the switch and global
zone to have access to the bootstrap network. Therefore we keep address
allocation and plumbing minimal for now.