[common] add variants to UpdateArtifactKind (#2300)
Add the list of updatable components from [RFD
300](https://rfd.shared.oxide.computer/rfd/0300#_list_of_updatable_components).

Remove "zone" since we determined that individual zones are not
updatable.

This list might change in the future, but it gives enough for me to work
with wrt the installinator and such.

## Questions

### Adding new artifact kinds

Adding these new update kinds was quite tricky for me, can it be made
simpler? (See
"Adding a new UpdateArtifactKind" in the PR.) Particularly concerned
about what seemed to be a circular dependency: the command listed
[here](https://github.com/oxidecomputer/omicron/blob/main/docs/adding-an-endpoint.adoc#5-testing)
failed to compile.

### Representing artifact kinds as strings

We have a bit of a conflict between the idiomatic ways to represent
artifact kinds:

* openapi-lint requires that enum variants have underscores in them
(e.g. `gimlet_sp`)
* on the command line, convention typically is to use dashes in them
(e.g. `gimlet-sp`).

There are a few ways to handle this:

1. Always use underscores everywhere, making the CLI somewhat less
idiomatic. This is what I've currently chosen, but I'm happy to change
this.
2. Use dashes in the `FromStr` and `Display` impls, and underscores in
the `Serialize`/`Deserialize` impls. This will work but I'm worried
about the discrepancy causing new bugs down the road.
3. Use underscores as the canonical representation, but add a separate
command-line parser which recognizes both underscores and dashes.
4. Always use dashes everywhere, silencing the OpenAPI lint somehow.
This seems to work with progenitor, but other client builders might have
trouble with this if they don't accept dashes.